requirements.txt
aiofiles==23.2.1
aiogram==2.25.1
aiohttp==3.8.0
aiosignal==1.3.1
aiosqlite==0.17.0
async-timeout==4.0.3
atlastk==0.13.3
attrs==23.2.0
Babel==2.9.1
certifi==2023.11.17
charset-normalizer==2.1.1
frozenlist==1.4.1
idna==3.6
iso8601==1.1.0
magic-filter==1.0.12
multidict==6.0.4
pypika-tortoise==0.1.6
pytz==2023.3.post1
tortoise-orm==0.20.0
typing_extensions==4.9.0
yarl==1.9.4
python-dotenv==1.0.1